% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_MainFun.R
\name{AsyAUC}
\alias{AsyAUC}
\title{Asymptotic functional diversity q profile by area under curve}
\usage{
AsyAUC(
  data,
  distM,
  datatype = "abundance",
  q = seq(0, 2, by = 0.25),
  nboot = 50,
  conf = 0.95,
  tau = NULL
)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances/incidences data.\cr
See \code{\link{AsyAUC}} for data details.}

\item{distM}{a pair wise distance matrix for all pairs of observed species in the pooled assemblage.\cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species by sampling-units incidence frequencies (\code{datatype = "incidence_freq"}), default is \code{"abundance"}. \cr}

\item{q}{a sequence of nonnegative integers specifying the diversity orders of FD. \cr
If \code{NULL},then \code{level} will be chosen as the minimum coverage of all sites after extrapolating each site to its double sample size. Default is \code{NULL}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications. Enter 0 to skip bootstrap;
in this case, the caculation of standard errors and confidence intervals will be skipped. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95. \cr}

\item{tau}{a sequence between 0 and 1 specifying tau for integrating area under curve. If \code{NULL}, \code{tau} = (0, 0.01, 0.02,..., 0.99, 1). Default is \code{NULL}.}
}
\value{
a table including the sample size, sample coverage,
method (Interpolated or Extrapolated), and diversity estimates with each \code{q} for the user-specified sample coverages. \cr\cr
}
\description{
\code{AsyAUC}: computes asymptotic functional diversity(FD) by area under curve thorough several thresholds.
}
\examples{
\donttest{
# Type (1) abundance data (treat incidence frequencies as abundances to save computation time.)
data(FunDdata.abu)
data <- FunDdata.abu$data
dij <-  FunDdata.abu$dij
out <- AsyAUC(data = data[,2], distM = dij, datatype = "abundance", nboot=0)
out
# Type (2) incidence frequency data 
data(FunDdata.inc)
data <- FunDdata.inc$data
dij <-  FunDdata.inc$dij
out <- AsyAUC(data = data, distM = dij, datatype = "incidence_freq", nboot=20)
out
}
}
