\name{Brazil}
\alias{Brazil}
\docType{data}
\title{
Brazil rainforests data (datatype = "abundance")
}
\description{
  The dataset includes tree species abundace counts in two assemblages/habitats (Edge and Interior) collected from
  rainforests in Brazil by Magnago et al. (2014, 2015, 2017). \cr
}
\usage{data(Brazil)}
\format{
  A list includes three objects: \cr
   $ data a data.frame with 425 tree species abundance data in two habitats: Edge and Interior. \cr
   $ tree a phylo object giving the phylogenetic tree in Newick format. \cr
   $ dist a symmetric distance matrix. \cr
 }
\references{
Magnago, L. F. S., Edwards, D. P., Edwards, F. A., Magrach, A., Martins, S. V., & Laurance, W. F. 2014. Functional attributes change but functional richness is unchanged after fragmentation of Brazilian Atlantic forests. Journal of Ecology 102: 475-485. \cr\cr
Magnago, L. F. S., Magrach, A., Barlow, J., Schaefer, C. E. G. R., Laurance, W. F., Martins, S. V., & Edwards, D. P. 2017. Do fragment size and edge effects predict carbon stocks in trees and lianas in tropical forests? Functional Ecology 31: 542-552. \cr\cr
Magnago, L. F. S., Rocha, M. F., Meyer, L., Martins, S. V., & Meira-Neto, J. A. A. 2015. Microclimatic conditions at forest edges have significant impacts on vegetation structure in large Atlantic forest fragments. Biodiversity and Conservation 24: 2305-2318. \cr\cr
}

\keyword{datasets}
