% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_MainFun.R
\name{FDinfo}
\alias{FDinfo}
\title{Exhibit basic data information}
\usage{
FDinfo(data, datatype)
}
\arguments{
\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).}

\item{x}{a vector/matrix/list of species abundances or incidence frequencies.\cr If \code{datatype = "incidence"}, 
then the first entry of the input data must be total number of sampling units, followed by species incidence frequencies.}
}
\value{
a data.frame of basic data information including sample size, observed species richness, sample coverage estimate, and the first ten abundance/incidence frequency counts.
}
\description{
\code{FDinfo}: exhibits basic data information
}
