% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{fish}
\alias{fish}
\title{Species-by-site raw incidence data}
\format{
A list includes three objects:
\describe{
  \item{$data}{a list with three species-by-site raw incidence records for 88 species and 60 sampling units.}
  \item{$tree}{a phylo object giving the phylogenetic tree in Newick format.}
  \item{$dist}{a symmetric distance matrix.}
}
}
\usage{
data(fish)
}
\description{
Species-by-site raw incidence data
}
\keyword{datasets}
