\name{Fish_incidence_data}
\alias{Fish_incidence_data}
\docType{data}
\title{
Species-by-site raw incidence data (datatype = "incidence_raw")
}
\description{
  Hinkley's fish data consists of the species-by-site raw incidence (detection/non-detection) matrix in two time periods (2013-2015, 2016-2018). \cr
}
\usage{data(Fish_incidence_data)}
\format{
  \code{Fish_incidence_data} is a list with two species-by-site raw incidence records for 88 species and 36 sampling units. Each matrix record the species as 0 (undetect) or 1 (detect) in each sampling units. \cr
  A list of 2: \cr
  $ 2013-2015: int [1:88, 1:36] 0 0 0 0 0 0 0 0 0 0 ... \cr
  $ 2016-2018: int [1:88, 1:36] 1 0 0 0 0 0 0 0 0 0 ... \cr
 }
 
\keyword{datasets}
