\name{Fish_phylo_tree}
\alias{Fish_phylo_tree}
\docType{data}
\title{
Phylogenetic tree for \code{Fish_incidence_data}
}
\description{
  This dataset includes a phylogenetic tree spanned by 88 species listed in the \code{Fish_incidence_data} dataset. \cr
}
\usage{data(Fish_phylo_tree)}
\format{
  \code{Fish_phylo_tree} is a list with the following phylogenetic information: \cr
  A list of 5: \cr
  $ edge: int [1:172, 1:2] 89 90 91 92 93 94 95 95 94 96 ... \cr
  $ edge.length: num [1:172] 0.023 0.057 0.023 0.851 0.011 0.023 0.011 0.011 0.023 0.011 ... \cr
  $ Nnode: int 85 \cr
  $ node.label: chr [1:85] "Vertebrata_-subphylum_in_Deuterostomia" ... \cr
  $ tip.label: chr [1:88] "Alosa_fallax" "Sardina_pilchardus" "Clupea_harengus" ... \cr
 }
\keyword{datasets}
