% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meta.R
\docType{data}
\name{FunDdata.abu}
\alias{FunDdata.abu}
\title{Abundance Data for Function FunD}
\format{
A list includes two objects:
\describe{
  \item{data}{is a data frame with 168 woody plants abundance data and two sites: Fushan and Kenting.}
  \item{dij}{is a trait-based functional distance between the i-th and j-th species among the 168 species.}
}
}
\source{
Carboni, M., Santoro, R., & Acosta, A. T. R. (2010). Are some communities of the coastal dune zonation more susceptible to alien plant invasion?. Journal of Plant Ecology, 3(2), 139-147.
}
\usage{
data(FunDdata.abu)
}
\description{
This data set describes the abundance and functional distance of 168 woody plants
in Taiwan dynamics plots as reported by Wu et al. (2010).It includes two
dynamic plots: Fushan at northern Taiwan and Kenting at southern Taiwan,
 which contains 94,86 woody distinct plants respectively.
}
\keyword{datasets}
