% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggAsy3D}
\alias{ggAsy3D}
\title{ggplot for Asymptotic diversity}
\usage{
ggAsy3D(outcome, profile = "q")
}
\arguments{
\item{outcome}{the outcome of the functions \code{Asy3D} .\cr}
}
\value{
a figure of asymptotic of empirical three-divrsity\cr\cr
}
\description{
\code{ggAsy3D} Plots q-profile, time-profile, and tau-profile based on the outcome of \code{Asy3D} using the ggplot2 package.\cr
It will only show the confidence interval of 'Estimated'.
}
\examples{
## example for abundance-based data
# diversity = 'TD'
data(spider)
out1 <- rbind(Asy3D(spider, diversity = 'TD', datatype = "abundance"), Obs3D(spider, diversity = 'TD', datatype = "abundance"))
ggAsy3D(out1)

# diversity = 'PD'
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out2 <- Asy3D(data, diversity = 'PD', datatype = "abundance", tree = tree, q = seq(0, 2, by = 0.25), nboot = 30, PDtype = "meanPD")
ggAsy3D(out2, profile = "q")

# diversity = 'FD' & FDtype = 'single'
data(FunDdata.abu)
data <- FunDdata.abu$data
dij <-  FunDdata.abu$dij
out3 <- Asy3D(data, diversity = 'FD', distM = dij, datatype = "abundance", q = c(0, 1, 2), threshold = seq(0, 0.6, 0.1), FDtype = 'single', nboot = 0)
ggAsy3D(out3, profile = "tau")

# diversity = 'FD' & FDtype = 'AUC'
data(FunDdata.abu)
data <- FunDdata.abu$data
dij <-  FunDdata.abu$dij
out4 <- Asy3D(data = data[,2], diversity = 'FD', distM = dij, datatype = "abundance", q = seq(0, 2, 0.5), nboot = 0)
ggAsy3D(out4)

## example for incidence-based data
# diversity = 'TD'
data(ant)
out5 <- rbind(Asy3D(ant, diversity = 'TD', datatype = "incidence_freq"), Obs3D(ant, diversity = 'TD', datatype = "incidence_freq"))
ggAsy3D(out5)

# diversity = 'PD'
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
nT <- data.inc$nT
out6 <- Asy3D(data, diversity = 'PD', nT = nT, datatype = "incidence_raw", tree = tree, q = c(0, 1, 2), reftime = seq(0.1, 82.8575, length.out = 40))
ggAsy3D(out6, profile = "time")

# diversity = 'FD' & FDtype = 'single'
data(FunDdata.inc)
data <- FunDdata.inc$data
dij <-  FunDdata.inc$dij
out7 <- Asy3D(data, diversity = 'FD', distM = dij, datatype = "incidence_freq", FDtype = 'single')
ggAsy3D(out7)

# diversity = 'FD' & FDtype = 'AUC'
data(FunDdata.inc)
data <- FunDdata.inc$data
dij <-  FunDdata.inc$dij
out8 <- Asy3D(data, diversity = 'FD', distM = dij, datatype = "incidence_freq", nboot = 20)
ggAsy3D(out8)

}
