% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_MainFun.R
\name{iNEXTAUC}
\alias{iNEXTAUC}
\title{Interpolation and extrapolation of functional diversity by area under curve}
\usage{
iNEXTAUC(
  data,
  distM,
  datatype = "abundance",
  q = c(0, 1, 2),
  endpoint = NULL,
  knots = 40,
  size = NULL,
  conf = 0.95,
  nboot = 50,
  tau = NULL
)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances/incidences data.\cr Type (1) abundance data: a S by N matrix/data.frame
where N is the number of assemblages. The element in i-th row and k-th is the abundance of species i in assemblage k. Please note
that the rownames of data must be the species names matching the species names in distance matrix and thus can't be empty.\cr
Type (2) incidence frequency data: the sampling unit is quadrat or transect, the observed species was only recorded as presence(detection)/absence(non-detection)
data in each sampling unit. Likewise, the rownames of data must be the species names matching the species names in phylogeny tree and thus can't be empty. \cr}

\item{distM}{a pair wise distance matrix for all pairs of observed species in the pooled assemblage.\cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species by sampling-units incidence frequencies (\code{datatype = "incidence_freq"}), default is \code{"abundance"}. \cr}

\item{q}{a sequence of nonnegative integers specifying the diversity orders of FD. Default is \code{c(0,1,2)}. \cr}

\item{endpoint}{an positive interger specifying the endpoint for rarefaction and
extrapolation range. If \code{NULL}, \code{endpoint} = double of the maximum reference sample size. It will be ignored if \code{size} is given. \cr}

\item{knots}{a positive integer specifying the number of knots between 1 and the \code{endpoint}. Default is 40.\cr}

\item{size}{a sequence of positive integers specifying the sample sizes for which FD estimates will be calculated. If \code{NULL}, then FD estimates will be
calculated for those sample sizes determined by the specified/default \code{endpoint} and \code{knots}. \cr}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95. \cr}

\item{nboot}{a positive integer specifying the number of bootstrap replications. Enter 0 to skip bootstrap;
in this case, the caculation of standard errors and confidence intervals will be skipped. Default is 50.}

\item{tau}{a sequence between 0 and 1 specifying tau for integrating area under curve. If \code{NULL}, \code{tau} = (0, 0.01, 0.02,..., 0.99, 1). Default is \code{NULL}.}
}
\value{
a table of AUC estimates and sample completeness for interpolated or extrapolated sample sizes along with their confidence intervals (if \code{nboot > 0}). \cr\cr
}
\description{
\code{iNEXTAUC}: the seamless rarefaction and extrapolation sampling curves of functional diversity(FD) by area under curve thorough several thresholds for q = 0, 1 and 2.
}
\examples{
\donttest{
# Type (1) abundance data (treat incidence frequencies as abundances to save computation time.)
data(FunDdata.abu)
data <- FunDdata.abu$data
dij <-  FunDdata.abu$dij
out <- iNEXTAUC(data = data[,1], distM = dij, datatype = "abundance", nboot = 0)
out
# Type (2) incidence frequency data 
data(FunDdata.inc)
data <- FunDdata.inc$data
dij <-  FunDdata.inc$dij
out <- iNEXTAUC(data = data, distM = dij, datatype = "incidence_freq", nboot = 20)
out
}
}
