% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_MainFun.R
\name{iNEXTFD}
\alias{iNEXTFD}
\title{Interpolation and extrapolation of functional diversity}
\usage{
iNEXTFD(
  data,
  distM,
  datatype = "abundance",
  q = c(0, 1, 2),
  endpoint = NULL,
  knots = 40,
  size = NULL,
  plot.type = 1:3,
  conf = 0.95,
  nboot = 50,
  threshold = NULL
)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances/incidences data.\cr Type (1) abundance data: a S by N matrix/data.frame
where N is the number of assemblages. The element in i-th row and k-th is the abundance of species i in assemblage k. Please note
that the rownames of data must be the species names matching the species names in distance matrix and thus can't be empty.\cr
Type (2) incidence frequency data: the sampling unit is quadrat or transect, the observed species was only recorded as presence(detection)/absence(non-detection)
data in each sampling unit. Likewise, the rownames of data must be the species names matching the species names in phylogeny tree and thus can't be empty. \cr}

\item{distM}{a pair wise distance matrix for all pairs of observed species in the pooled assemblage.\cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species by sampling-units incidence frequencies (\code{datatype = "incidence_freq"}), default is \code{"abundance"}. \cr}

\item{q}{a sequence of nonnegative integers specifying the diversity orders of FD. Default is \code{c(0,1,2)}. \cr}

\item{endpoint}{an positive interger specifying the endpoint for rarefaction and
extrapolation range. If \code{NULL}, \code{endpoint} = double of the maximum reference sample size. It will be ignored if \code{size} is given. \cr}

\item{knots}{a positive integer specifying the number of knots between 1 and the \code{endpoint}. Default is 40.\cr}

\item{size}{a sequence of positive integers specifying the sample sizes for which FD estimates will be calculated. If \code{NULL}, then FD estimates will be
calculated for those sample sizes determined by the specified/default \code{endpoint} and \code{knots}. \cr}

\item{plot.type}{a positive integer vector specifying types of curves. Three types of plots: sample-size-based rarefaction and extrapolation curve (\code{plot.type = 1});
coverage-based rarefaction and extrapolation curve (\code{plot.type = 2}); sample completeness curve (\code{plot.type = 3}). Default is \code{c(1,2,3)}. \cr}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95. \cr}

\item{nboot}{a positive integer specifying the number of bootstrap replications. Enter 0 to skip bootstrap;
in this case, the caculation of standard errors and confidence intervals will be skipped. Default is 50.}

\item{threshold}{a sequence between 0 and 1 specifying tau. If \code{NULL}, \code{threshold} = dmean/2. Default is \code{NULL}.}
}
\value{
a table of FD estimates and sample completeness for interpolated or extrapolated sample sizes along with their confidence intervals (if \code{nboot > 0}). \cr\cr
}
\description{
\code{iNEXTFD}: the seamless rarefaction and extrapolation sampling curves of functional diversity(FD) for q = 0, 1 and 2.
See Chao et al. (2019) for pertinent background and methods.
}
\examples{
\donttest{
# Type (1) abundance data (treat incidence frequencies as abundances to save computation time.)
data(FunDdata.abu)
data <- FunDdata.abu$data
dij <-  FunDdata.abu$dij
out <- iNEXTFD(data = data, distM = dij,datatype = "abundance", nboot = 0)
out
# Type (2) incidence frequency data 
data(FunDdata.inc)
data <- FunDdata.inc$data
dij <-  FunDdata.inc$dij
out <- iNEXTFD(data = data, distM = dij,datatype = "incidence_freq")
out
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). functional diversity measures based on Hill numbers. Philosophical Transactions of the Royal Society B., 365, 3599-3609.\cr\cr
Chao, A., Chiu, C.-H., Hsieh, T. C., Davis, T., Nipperess, D., and Faith, D. (2015) Rarefaction and extrapolation of functional diversity. Methods in Ecology and Evolution, 6, 380-388.\cr\cr
Hsieh, T. C. and Chao, A. (2017). Rarefaction and extrapolation: making fair comparison of abundance-sensitive functional diversity among multiple assemblages. Systematic Biology 66, 100-111.
}
