% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFun.R
\name{print.iNEXT.3D}
\alias{print.iNEXT.3D}
\title{Printing iNEXT.3D object}
\usage{
\method{print}{iNEXT.3D}(x, ...)
}
\arguments{
\item{x}{an \code{iNEXT.3D} object computed by \code{\link{iNEXT3D}}.}

\item{...}{additional arguments.}
}
\description{
\code{print.iNEXT.3D}: Print method for objects inheriting from class "iNEXT.3D"
}
