% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{print.iNEXT3D}
\alias{print.iNEXT3D}
\title{Printing iNEXT.3D object}
\usage{
\method{print}{iNEXT3D}(x, ...)
}
\arguments{
\item{x}{an \code{iNEXT3D} object computed by \code{\link{iNEXT3D}}.}

\item{...}{additional arguments.}
}
\description{
\code{print.iNEXT3D}: Print method for objects inheriting from class "iNEXT3D"
}
