% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{Asy3D}
\alias{Asy3D}
\title{Asymptotic diversity q profile}
\usage{
Asy3D(
  data,
  diversity = "TD",
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 50,
  conf = 0.95,
  tree,
  nT,
  reftime = NULL,
  PDtype = "PD",
  distM,
  FDtype = "AUC",
  threshold = NULL
)
}
\arguments{
\item{data}{a matrix/data.frame (species by sites), or list of species abundances or incidence frequencies. If \code{datatype = "incidence_freq"}, then the first entry of the input data must be total number of sampling units in each column or list.}

\item{diversity}{a choice of three-level diversity: 'TD' = 'Taxonomic', 'PD' = 'Phylogenetic', and 'FD' = 'Functional'.}

\item{q}{a nonnegative value or sequence specifying the diversity order. Default is seq(0, 2, by = 0.2).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species-by-site incidence frequencies data (\code{datatype = "incidence_freq"}). Default is "abundance".}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}

\item{tree}{a phylo object describing the phylogenetic tree in Newick format for all observed species in the pooled assemblage. It is necessary when \code{diversity = 'PD'}.}

\item{nT}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the number of sampling units in each assemblage.
If \code{names(nT) = NULL}, then assemblage are automatically named as "assemblage1", "assemblage2",..., etc.
It is necessary when \code{diversity = 'PD'} and \code{datatype = "incidence_raw"}.}

\item{reftime}{is set to be the tree depth of the phylogenetic tree, which is spanned by all the observed species in
the pooled assemblage. Default is \code{NULL}. It will be use when \code{diversity = 'PD'}.}

\item{PDtype}{desired phylogenetic diversity type: \code{PDtype = "PD"} for Chao et al. (2010) phylogenetic diversity
and \code{PDtype = "meanPD"} for mean phylogenetic diversity (phylogenetic Hill number). It will be use when \code{diversity = 'PD'}. Default is \code{"PD"}.}

\item{distM}{a pair wise distance matrix for all pairs of observed species in the pooled assemblage. It will be use when \code{diversity = 'FD'}.}

\item{FDtype}{a binary selection for functional type. \code{FDtype = "single"} computes diversity under certain threshold. \code{FDtype = "AUC"} computes diversity which 
integrates several threshold between zero and one to get diversity. Default is \code{"AUC"}}

\item{threshold}{a sequence between 0 and 1 specifying tau. If \code{NULL}, \code{threshold = } dmean. Default is \code{NULL}. It will be use when \code{diversity = 'FD'}.}
}
\value{
a table of diversity q profile by 'Estimated'.
}
\description{
\code{Asy3D} The estimated diversity of order q
}
\examples{
## example for abundance-based data
# diversity = 'TD'
data(spider)
out1 <- Asy3D(spider, diversity = 'TD', datatype = "abundance")
out1

# diversity = 'PD'
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out2 <- Asy3D(data, diversity = 'PD', datatype = "abundance", tree = tree, q = seq(0, 2, by = 0.25), nboot = 30)
out2

# diversity = 'FD' & FDtype = 'single'
data(FunDdata.abu)
data <- FunDdata.abu$data
dij <-  FunDdata.abu$dij
out3 <- Asy3D(data, diversity = 'FD', distM = dij, datatype = "abundance", q = seq(0, 2, 0.5), FDtype = 'single', nboot = 0)
out3

# diversity = 'FD' & FDtype = 'AUC'
data(FunDdata.abu)
data <- FunDdata.abu$data
dij <-  FunDdata.abu$dij
out4 <- Asy3D(data = data[,2], diversity = 'FD', distM = dij, datatype = "abundance", q = seq(0, 2, 0.5), nboot = 0)
out4

## example for incidence-based data
# diversity = 'TD'
data(ant)
out5 <- Asy3D(ant, diversity = 'TD', datatype = "incidence_freq")
out5

# diversity = 'PD'
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
nT <- data.inc$nT
out6 <- Asy3D(data, diversity = 'PD', nT = nT, datatype = "incidence_raw", tree = tree, q = seq(0, 2, by = 0.25))
out6

# diversity = 'FD' & FDtype = 'single'
data(FunDdata.inc)
data <- FunDdata.inc$data
dij <-  FunDdata.inc$dij
out7 <- Asy3D(data, diversity = 'FD', distM = dij, datatype = "incidence_freq", FDtype = 'single')
out7

# diversity = 'FD' & FDtype = 'AUC'
data(FunDdata.inc)
data <- FunDdata.inc$data
dij <-  FunDdata.inc$dij
out8 <- Asy3D(data, diversity = 'FD', distM = dij, datatype = "incidence_freq", nboot = 20)
out8

}
