% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{DataInfo3D}
\alias{DataInfo3D}
\title{Exhibit basic data information for three diversity class}
\usage{
DataInfo3D(
  data,
  diversity,
  datatype,
  tree = NULL,
  nT = NULL,
  reftime = NULL,
  distM = NULL,
  FDtype = "AUC",
  threshold = NULL
)
}
\arguments{
\item{data}{a matrix, data.frame (species by sites if \code{datatype = "abundance" or "incidence_freq"}, 
species by subplot if \code {datatype = "incidence_raw"}), or list of species abundances or incidence frequencies/raw. If \code{datatype = "incidence_freq"}, then the first entry of the input data must be total number of sampling units in each column or list.}

\item{diversity}{a choice of three-level diversity: 'TD' = 'Taxonomic', 'PD' = 'Phylogenetic', and 'FD' = 'Functional' under certain threshold. Besides,'AUC' is the fourth choice which 
integrates several threshold functional diversity to get diversity.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).}

\item{tree}{a phylo object describing the phylogenetic tree in Newick format for all observed species in the pooled assemblage. It is necessary when \code{class = 'PD'}.}

\item{nT}{needed only for the matrix or data.frame which \code{datatype = "incidence_raw"} , a sequence of named nonnegative integers specifying the number of sampling units in each assemblage.}

\item{distM}{a pair wise distance matrix for all pairs of observed species in the pooled assemblage. It will be use when \code{class = 'FD' or 'AUC'}.}

\item{threshold}{a sequence between 0 and 1 specifying tau. If \code{NULL}, \code{threshold = } dmean. Default is \code{NULL}. It will be use when \code{class = 'FD'}.}
}
\value{
a data.frame or lists of objects containing data.frame (depending on the number of diversity class required) for summarizing data information.
}
\description{
\code{dataInfo}: Exhibit basic data information for three diversity class
}
\examples{
## example for abundance based data (matrix)
# diversity = 'TD'
data(beetles)
DataInfo3D(data =beetles.abu$data,"TD", "abundance")

# diversity = 'PD'
data(beetles)
DataInfo3D(data =beetles.abu$data,"PD", "abundance", tree = beetles.abu$tree)

# diversity = 'FD'
data(beetles)
DataInfo3D(data = beetles.abu$data, diversity = "FD", datatype = "abundance", distM = beetles.abu$distance.matrix)

# multiple diversity classes required
DataInfo3D(data = beetles.abu$data, diversity = c("TD","PD","FD"), datatype = "abundance", tree = beetles.abu$tree, distM = beetles.abu$distance.matrix)

## example for incidence-freq datatype (matrix of incidence frequency, 'datatype' = 'incidence_freq' is not accepted in 'class' = 'PD')  
# diversity = c('TD', 'FD')
data(fish)
DataInfo3D(data = fish$incidence_freq, diversity = c('TD', 'FD'), datatype = "incidence_freq", distM = fish$dis)

## example for incidence-raw datatype (lists of incidence-raw data)
# diversity = c('TD', 'PD','FD')
data(fish)
data(fish.raw)
DataInfo3D(data = fish.raw, diversity = c('TD', 'PD','FD'), datatype = "incidence_raw", distM = fish$dis)

}
