% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{DataInfo3D}
\alias{DataInfo3D}
\title{Data information for reference samples}
\usage{
DataInfo3D(
  data,
  diversity = "TD",
  datatype = "abundance",
  nT = NULL,
  PDtree,
  PDreftime = NULL,
  FDdistM,
  FDtype = "AUC",
  FDtau = NULL
)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} (species by sites), or \code{list} of species abundance/incidence frequencies.\cr 
If \code{datatype = "incidence_freq"}, then the first entry of the input data must be total number of sampling units in each column or list.}

\item{diversity}{selection of diversity type: 'TD' = Taxonomic diversity, 'PD' = Phylogenetic diversity, and 'FD' = Functional diversity.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}), sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).}

\item{nT}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the number of sampling units in each assemblage. \cr
If \code{names(nT) = NULL}, then assemblage are automatically named as "assemblage1", "assemblage2",..., etc. \cr
It is necessary when \code{diversity = 'PD'} and \code{datatype = "incidence_raw"}.}

\item{PDtree}{(required argument only for \code{diversity = "PD"}), a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{PDreftime}{(argument only for \code{diversity = "PD"}), a vector of numerical values specifying reference times for PD. Default is \code{NULL} (i.e., the age of the root of PDtree).}

\item{FDdistM}{(required argument only for \code{diversity = "FD"}), a species pairwise distance matrix for all species in the pooled assemblage.}

\item{FDtype}{(argument only for \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_values"} for FD under specified threshold values, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(argument only for \code{diversity = "FD"} and \code{FDtype = "tau_values"}), a numerical vector between 0 and 1 specifying tau values (threshold levels). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}
}
\value{
a data.frame including basic data information.\cr\cr 
For abundance data, basic information shared by TD, mean-PD and FD
 includes assemblage name (Assemblage),
sample size (n), observed species richness (S.obs), sample coverage estimates of the reference sample (SC(n)), 
sample coverage estimate for twice the reference sample size (SC(2n)). Other additional information is given below.\cr\cr
(1) TD: the first five species abundance (f1-f10).\cr\cr
(2) Mean-PD: the the observed total branch length in the phylogenetic tree (PD.obs), 
the number of singletons (f1*) and doubletons (f2*) in the node/branch abundance set, as well as the total branch length 
of those singletons (g1) and of those doubletons (g2), and the reference time (Reftime).\cr\cr
(3) FD (\code{FDtype = "AUC"}): the minimum distance among all non-diagonal elements in the distance matrix (dmin), the mean distance
(dmean), and the maximum distance (dmax) in the distance matrix.\cr \cr
(4) FD (\code{FDtype = "tau_value"}): the number of singletons (a1*) and of doubletons (a2*) among the functionally indistinct
 set at the specified threshold level 'Tau', as well as the total contribution of singletons (h1) and of doubletons (h2)
  at the specified threshold level 'Tau'.\cr\cr
 
 For incidence data, the basic information for TD includes assemblage name (Assemblage), number of sampling units (T), 
 total number of incidences (U), observed species richness (S.obs), 
 sample coverage estimates of the reference sample (SC(T)), sample coverage estimate for twice the reference sample size
 (SC(2T)), as well as the first ten species incidence frequency counts (Q1-Q10). For mean-PD and FD, output is similar to that
 for abundance data.
}
\description{
\code{DataInfo3D} Provides basic data information
}
\examples{
# diversity = 'TD' for abundance-based data
data(Brazil)
DataInfo3D(Brazil$data, diversity = 'TD', datatype = "abundance")


# diversity = 'PD' for abundance-based data
data(Brazil)
data <- Brazil$data
tree <- Brazil$tree
DataInfo3D(data, diversity = 'PD', datatype = "abundance", PDtree = tree)


# diversity = 'FD' & FDtype = 'tau_values' for abundance-based data
data(Brazil)
data <- Brazil$data
distM <- Brazil$dist
DataInfo3D(data, diversity = 'FD', datatype = "abundance", FDdistM = distM, FDtype = 'tau_values')


# diversity = 'FD' & FDtype = 'AUC' for abundance-based data
data(Brazil)
data <- Brazil$data
distM <- Brazil$dist
DataInfo3D(data, diversity = 'FD', datatype = "abundance", FDdistM = distM, FDtype = 'AUC')


# diversity = 'TD' for incidence-based data
data(fish)
DataInfo3D(fish$data, diversity = 'TD', datatype = "incidence_raw")


# diversity = 'PD' for incidence-based data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
nT <- data.inc$nT
DataInfo3D(data, diversity = 'PD', datatype = "incidence_raw", nT = nT, PDtree = tree)


# diversity = 'FD' & FDtype = 'tau_values' for incidence-based data
data(fish)
data <- fish$data
distM <- fish$dist
DataInfo3D(data, diversity = 'FD', datatype = "incidence_raw", 
           FDdistM = distM, FDtype = 'tau_values')


# diversity = 'FD' & FDtype = 'AUC' for incidence-based data
data(fish)
data <- fish$data
distM <- fish$dist
DataInfo3D(data, diversity = 'FD', datatype = "incidence_raw", FDdistM = distM, FDtype = 'AUC')


}
