% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{estimate3D}
\alias{estimate3D}
\title{Compute species diversity with a particular of sample size/coverage}
\usage{
estimate3D(
  data,
  diversity = "TD",
  q = c(0, 1, 2),
  datatype = "abundance",
  base = "coverage",
  level = NULL,
  nboot = 50,
  conf = 0.95,
  nT,
  PDtree,
  PDreftime = NULL,
  PDtype = "meanPD",
  FDdistM,
  FDtype = "AUC",
  FDtau = NULL
)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} (species by sites), or \code{list} of species abundance/incidence frequencies.\cr 
If \code{datatype = "incidence_freq"}, then the first entry of the input data must be total number of sampling units in each column or list.}

\item{diversity}{selection of diversity type: 'TD' = Taxonomic diversity, 'PD' = Phylogenetic diversity, and 'FD' = Functional diversity.}

\item{q}{a numerical vector specifying the diversity orders.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}), sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).}

\item{base}{Selection of sample-size-based (\code{base="size"}) or coverage-based (\code{base="coverage"}) rarefaction and extrapolation.}

\item{level}{a sequence specifying the particular sample sizes or sample coverages (between 0 and 1). \cr
If \code{base="size"} and \code{level=NULL}, then this function computes the diversity estimates for the minimum sample size among all sites extrapolated to double reference sizes.  \cr
If \code{base="coverage"} and \code{level=NULL}, then this function computes the diversity estimates for the minimum sample coverage among all sites extrapolated to double reference sizes.}

\item{nboot}{an integer specifying the number of replications. If confidence interval is not desired, use 0 to skip this time-consuming step.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95.}

\item{nT}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the number of sampling units in each assemblage. \cr
If \code{names(nT) = NULL}, then assemblage are automatically named as "assemblage1", "assemblage2",..., etc. \cr
It is necessary when \code{diversity = 'PD'} and \code{datatype = "incidence_raw"}.}

\item{PDtree}{a phylo object describing the phylogenetic tree in Newick format for all observed species in the pooled assemblage. It is necessary when \code{diversity = 'PD'}.}

\item{PDreftime}{Select several reference time points for \code{diversity = 'PD'}. Default is NULL.}

\item{PDtype}{Select phylogenetic diversity type: \code{PDtype = "PD"} for Chao et al. (2010) phylogenetic diversity and \code{PDtype = "meanPD"} for mean phylogenetic diversity (phylogenetic Hill number). It will be use when \code{diversity = 'PD'}. Default is \code{"meanPD"}.}

\item{FDdistM}{a pair wise distance matrix for all pairs of observed species in the pooled assemblage. It will be use when \code{diversity = 'FD'}.}

\item{FDtype}{a binary selection for FD. \code{FDtype = "tau_values"} computes diversity under certain threshold values. \code{FDtype = "AUC"} computes an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{a sequence between 0 and 1 specifying tau. If \code{NULL}, \code{threshold = } dmean. Default is \code{NULL}. It will be use when \code{diversity = 'FD'} and \code{FDtype = "tau_values"}.}
}
\value{
a \code{data.frame} of species diversity table including the sample size, sample coverage,
method (rarefaction or extrapolation), and diversity estimates with q = 0, 1, and 2 for the user-specified sample size or sample coverage.
}
\description{
\code{estimate3D}: computes species diversity (Hill numbers with q = 0, 1 and 2) with a particular user-specified level of sample size or sample coverage.
}
\examples{
# diversity = 'TD'
data(spider)
out1 <- estimate3D(spider, diversity = 'TD', q = c(0,1,2), datatype = "abundance", base = "size")
out1

# diversity = 'PD'
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out2 <- estimate3D(data, diversity = 'PD', datatype = "abundance", base = "coverage", PDtree = tree)
out2

# diversity = 'FD' & FDtype = 'tau_values'
data(FunDdata.abu)
data <- FunDdata.abu$data
distM <-  FunDdata.abu$dij
out3 <- estimate3D(data, diversity = 'FD', datatype = "abundance", base = "size", FDdistM = distM, FDtype = 'tau_values')
out3

# diversity = 'FD' & FDtype = 'AUC'
data(FunDdata.abu)
data <- FunDdata.abu$data
distM <-  FunDdata.abu$dij
out4 <- estimate3D(data = data[,2], diversity = 'FD', datatype = "abundance", base = "coverage", nboot = 0, FDdistM = distM)
out4

## example for incidence-based data
# diversity = 'TD'
data(ant)
out5 <- estimate3D(ant, diversity = 'TD', q = c(0,1,2), datatype = "incidence_freq", base = "coverage", level=0.985)
out5

# diversity = 'PD'
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
nT <- data.inc$nT
out6 <- estimate3D(data, diversity = 'PD', datatype = "incidence_raw", base = "size", nT = nT, PDtree = tree)
out6

# diversity = 'FD' & FDtype = 'tau_values'
data(FunDdata.inc)
data <- FunDdata.inc$data
distM <-  FunDdata.inc$dij
out7 <- estimate3D(data, diversity = 'FD', datatype = "incidence_freq", base = "coverage", FDdistM = distM, FDtype = 'tau_values')
out7

# diversity = 'FD' & FDtype = 'AUC'
data(FunDdata.inc)
data <- FunDdata.inc$data
distM <-  FunDdata.inc$dij
out8 <- estimate3D(data, diversity = 'FD', datatype = "incidence_freq", base = "size", nboot = 20, FDdistM = distM)
out8

}
