% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{estimate3D}
\alias{estimate3D}
\title{Compute diversity with a particular of sample size/coverage}
\usage{
estimate3D(
  data,
  diversity = "TD",
  q = c(0, 1, 2),
  datatype = "abundance",
  base = "coverage",
  level = NULL,
  nboot = 50,
  conf = 0.95,
  nT = NULL,
  PDtree,
  PDreftime = NULL,
  PDtype = "meanPD",
  FDdistM,
  FDtype = "AUC",
  FDtau = NULL,
  FDcut_number = 50
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a vector of species abundances (for a single assemblage), matrix/data.frame (species by assemblages), or a list of species abundance vectors. \cr
(b) For \code{datatype = "incidence_raw"}, data can be input as a list of matrix/data.frame (species by sampling units); data can also be input as a matrix/data.frame by merging all sampling units across assemblages based on species identity; in this case, the number of sampling units (nT, see below) must be input.}

\item{diversity}{selection of diversity type: \code{'TD'} = Taxonomic diversity, \code{'PD'} = Phylogenetic diversity, and \code{'FD'} = Functional diversity.}

\item{q}{a numerical vector specifying the diversity orders. Default is c(0, 1, 2).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being 0 (non-detection) or 1 (detection).}

\item{base}{selection of sample-size-based (\code{base = "size"}) or coverage-based (\code{base = "coverage"}) rarefaction and extrapolation.}

\item{level}{A numerical vector specifying the particular sample sizes or sample coverages (between 0 and 1). \cr
If \code{base = "coverage"} (default) and \code{level = NULL}, then this function computes the diversity estimates for the minimum sample coverage among all samples extrapolated to double reference sizes. \cr
If \code{base = "size"} and \code{level = NULL}, then this function computes the diversity estimates for the minimum sample size among all samples extrapolated to double reference sizes.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}

\item{nT}{(required only when \code{datatype = "incidence_raw"} and input data is matrix/data.frame) a vector of nonnegative integers specifying the number of sampling units in each assemblage. If assemblage names are not specified, then assemblages are automatically named as "assemblage1", "assemblage2",..., etc.}

\item{PDtree}{(required argument only for \code{diversity = "PD"}), a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{PDreftime}{(argument only for \code{diversity = "PD"}), a vector of numerical values specifying reference times for PD. Default is \code{NULL} (i.e., the age of the root of PDtree).}

\item{PDtype}{(argument only for \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"} (effective total branch length) or \code{PDtype = "meanPD"} (effective number of equally divergent lineages). Default is \code{"meanPD"}, where \code{meanPD = PD/tree depth}.}

\item{FDdistM}{(required argument only for \code{diversity = "FD"}), a species pairwise distance matrix for all species in the pooled assemblage.}

\item{FDtype}{(argument only for \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_values"} for FD under specified threshold values, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(argument only for \code{diversity = "FD"} and \code{FDtype = "tau_values"}), a numerical vector between 0 and 1 specifying tau values (threshold levels). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}

\item{FDcut_number}{(argument only for \code{diversity = "FD"} and \code{FDtype = "AUC"}), a numeric number to cut [0, 1] interval into equal-spaced sub-intervals to obtain the AUC value by integrating the tau-profile. Equivalently, the number of tau values that will be considered to compute the integrated AUC value. Default is FDcut_number = 50 A larger value can be set to obtain more accurate AUC value.}
}
\value{
a \code{data.frame} of diversity table including the following arguments: (when \code{base = "coverage"})
\item{Assemblage}{the assemblage name}
\item{Order.q}{the diversity order of q.}
\item{SC}{the target standardized coverage value.}
\item{m, nT}{the corresponding sample size (or sampling units) for the standardized coverage value.}
\item{qTD, qPD, qFD}{the estimated diversity of order q for the target coverage value. The estimate for complete coverage (or size = infinity) represents the estimated asymptotic diversity.}
\item{Method}{Rarefaction, Observed, or Extrapolation, depending on whether the target coverage is less than, equal to, or greater than the coverage of the reference sample.}
\item{s.e.}{standard error of diversity estimate.}
\item{qTD.LCL, qPD.LCL, qFD.LCL and qTD.UCL, qPD.UCL, qFD.UCL}{the bootstrap lower and upper confidence limits for the diversity of order q at the specified level. (with a default value of 0.95)}
\item{Reftime}{reference times for PD.}
\item{Type}{"PD" (effective total branch length) or "meanPD" (effective number of equally divergent lineages) for PD.}
\item{Tau}{the threshold of functional distinctiveness between any two species for FD (under FDtype = tau_values).}
Similar output is obtained for \code{base = "size"}. \cr\cr
}
\description{
\code{estimate3D}: computes diversity (Hill-Chao number with q = 0, 1 and 2) with a particular user-specified level of sample size or sample coverage.
}
\examples{
# diversity = 'TD' for abundance-based data
data(Brazil_rainforest_data)
output1 <- estimate3D(Brazil_rainforest_data$data, diversity = 'TD', q = c(0,1,2), datatype = "abundance", base = "size")
output1


# diversity = 'PD' for abundance-based data
data(Brazil_rainforest_data)
data <- Brazil_rainforest_data$data
tree <- Brazil_rainforest_data$tree
output2 <- estimate3D(data, diversity = 'PD', datatype = "abundance", base = "coverage", PDtree = tree)
output2


# diversity = 'FD' & FDtype = 'tau_values' for abundance-based data
data(Brazil_rainforest_data)
data <- Brazil_rainforest_data$data
distM <- Brazil_rainforest_data$dist
output3 <- estimate3D(data, diversity = 'FD', datatype = "abundance", base = "size", 
                      FDdistM = distM, FDtype = 'tau_values')
output3


# diversity = 'FD' & FDtype = 'AUC' for abundance-based data
data(Brazil_rainforest_data)
data <- Brazil_rainforest_data$data
distM <- Brazil_rainforest_data$dist
output4 <- estimate3D(data, diversity = 'FD', datatype = "abundance", base = "coverage", 
                      nboot = 20, FDdistM = distM)
output4


# diversity = 'TD' for incidence-based data
data(fish_incidence_data)
output5 <- estimate3D(fish_incidence_data$data, diversity = 'TD', q = c(0,1,2), datatype = "incidence_raw", 
                      base = "coverage")
output5


# diversity = 'PD' for incidence-based data
data(fish_incidence_data)
data <- fish_incidence_data$data
tree <- fish_incidence_data$tree
output6 <- estimate3D(data, diversity = 'PD', datatype = "incidence_raw", 
                      base = "size", PDtree = tree)
output6


# diversity = 'FD' & FDtype = 'tau_values' for incidence-based data
data(fish_incidence_data)
data <- fish_incidence_data$data
distM <- fish_incidence_data$dist
output7 <- estimate3D(data, diversity = 'FD', datatype = "incidence_raw", base = "coverage", 
                      FDdistM = distM, FDtype = 'tau_values')
output7


# diversity = 'FD' & FDtype = 'AUC' for incidence-based data
data(fish_incidence_data)
data <- fish_incidence_data$data
distM <- fish_incidence_data$dist
output8 <- estimate3D(data, diversity = 'FD', datatype = "incidence_raw", base = "size", 
                      nboot = 20, FDdistM = distM)
output8


}
\references{
Chao, A., Henderson, P. A., Chiu, C.-H., Moyes, F., Hu, K.-H., Dornelas, M and Magurran, A. E. (2021). Measuring temporal change in alpha diversity: a framework integrating taxonomic, phylogenetic and functional diversity and the iNEXT.3D standardization. Methods in Ecology and Evolution, 12, 1926-1940.
}
