\name{fish_incidence_data}
\alias{fish_incidence_data}
\docType{data}
\title{
Species-by-site raw incidence data (datatype = "incidence_raw")
}
\description{
  Hinkley's fish data consists of the species-by-site raw incidence (detection/non-detection) matrix in two time periods (2013-2015, 2016-2018). Besides, it includes a phylogenetic tree and a species pairwise distance matrix.\cr
}
\usage{data(fish_incidence_data)}
\format{
  A list includes three objects: \cr
   $ data: a list with two species-by-site raw incidence records for 88 species and 36 sampling units. \cr
   $ tree: a phylo object giving the phylogenetic tree in Newick format. \cr
   $ dist: a symmetric distance matrix. \cr
 }
\keyword{datasets}
