% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggAO3D}
\alias{ggAO3D}
\title{ggplot for Asymptotic diversity}
\usage{
ggAO3D(outcome, profile = "q")
}
\arguments{
\item{outcome}{the outcome of the functions \code{AO3D} or \code{obs3D}.\cr}

\item{profile}{a selection of profile versus to diversity. User can choose \code{'q'}, \code{'time'}, and \code{'tau'}. Default is \code{'q'} profile. \code{'time'} profile for only when \code{diversity = "PD"}. \code{'tau'} profile for only when \code{diversity = "FD"} and \code{FDtype = "tau_values"}.\cr}
}
\value{
a figure of asymptotic of empirical three-divrsity\cr\cr
}
\description{
\code{ggAO3D} Plots q-profile, time-profile, and tau-profile based on the outcome of \code{AO3D} using the ggplot2 package.\cr
It will only show the confidence interval of 'Estimated'.
}
\examples{
## example for abundance-based data
# diversity = 'TD'
data(dunes)
out1 <- AO3D(dunes$data, diversity = 'TD', datatype = "abundance", method = c('Estimated', 'Empirical'))
ggAO3D(out1)

# diversity = 'PD'
data(dunes)
data <- dunes$data
tree <- dunes$tree
out2 <- AO3D(data, diversity = 'PD', q = seq(0, 2, by = 0.25), datatype = "abundance", nboot = 30, method = c('Estimated', 'Empirical'), PDtree = tree, PDtype = "meanPD")
ggAO3D(out2, profile = "q")

# diversity = 'FD' & FDtype = 'tau_values'
data(dunes)
data <- dunes$data
distM <-  dunes$dist
out3 <- AO3D(data, diversity = 'FD', q = c(0, 1, 2), datatype = "abundance", nboot = 0, method = c('Estimated', 'Empirical'), FDtau = seq(0, 0.6, 0.1), FDdistM = distM, FDtype = 'tau_values')
ggAO3D(out3, profile = "tau")

# diversity = 'FD' & FDtype = 'AUC'
data(dunes)
data <- dunes$data
distM <- dunes$dist
out4 <- AO3D(data, diversity = 'FD', q = seq(0, 2, 0.5), datatype = "abundance", nboot = 0, method = c('Estimated', 'Empirical'), FDdistM = distM)
ggAO3D(out4)

## example for incidence-based data
# diversity = 'TD'
data(fish)
out5 <- AO3D(fish$data, diversity = 'TD', datatype = "incidence_raw", method = c('Estimated', 'Empirical'))
ggAO3D(out5)

# diversity = 'PD'
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
nT <- data.inc$nT
out6 <- AO3D(data, diversity = 'PD', q = c(0, 1, 2), datatype = "incidence_raw", nT = nT, method = c('Estimated', 'Empirical'), PDtree = tree, PDreftime = seq(0.1, 82.8575, length.out = 40))
ggAO3D(out6, profile = "time")

# diversity = 'FD' & FDtype = 'tau_values'
data(fish)
data <- fish$data
distM <- fish$dist
out7 <- AO3D(data, diversity = 'FD', datatype = "incidence_raw", method = c('Estimated', 'Empirical'), FDdistM = distM, FDtype = 'tau_values')
ggAO3D(out7)

# diversity = 'FD' & FDtype = 'AUC'
data(fish)
data <- fish$data
distM <- fish$dist
out8 <- AO3D(data, diversity = 'FD', datatype = "incidence_raw", nboot = 20, method = c('Estimated', 'Empirical'), FDdistM = distM)
ggAO3D(out8)

}
