% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggiNEXT3D}
\alias{ggiNEXT3D}
\title{ggplot2 extension for an iNEXT object}
\usage{
ggiNEXT3D(outcome, type = 1:3, facet.var = "Assemblage", color.var = "Order.q")
}
\arguments{
\item{type}{three types of plots: sample-size-based rarefaction/extrapolation curve (\code{type = 1}); 
sample completeness curve (\code{type = 2}); coverage-based rarefaction/extrapolation curve (\code{type = 3}).}

\item{facet.var}{create a separate plot for each value of a specified variable: 
no separation \cr (\code{facet.var="None"}); 
a separate plot for each diversity order (\code{facet.var="Order.q"}); 
a separate plot for each assemblage (\code{facet.var="Assemblage"}); 
a separate plot for each combination of order x assemblage (\code{facet.var="Both"}).}

\item{color.var}{create curves in different colors for values of a specified variable:
all curves are in the same color (\code{color.var="None"}); 
use different colors for diversity orders (\code{color.var="Order.q"}); 
use different colors for sites (\code{color.var="Assemblage"}); 
use different colors for combinations of order x assemblage (\code{color.var="Both"}).}

\item{x}{an \code{iNEXT} object computed by \code{\link{iNEXT}}.}

\item{...}{other arguments passed on to methods. Not currently used.}
}
\value{
a ggplot2 object
}
\description{
\code{ggiNEXT3D}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{iNEXT3D}} Object to plot sample-size- and coverage-based rarefaction/extrapolation curves along with a bridging sample completeness curve
}
\examples{
## example for abundance based data (list of vector)
# diversity = 'TD'
data(spider)
out1 <- iNEXT3D(spider, diversity = 'TD', q = c(0,1,2), datatype = "abundance")
ggiNEXT3D(out1, facet.var = "Assemblage")

# diversity = 'PD'
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out2 <- iNEXT3D(data, diversity = 'PD', q = c(0,1,2), datatype = "abundance", nboot = 30, PDtree = tree)
ggiNEXT3D(out2, type = c(1, 3))

# diversity = 'FD' & FDtype = 'tau_values'
data(FunDdata.abu)
data <- FunDdata.abu$data
distM <-  FunDdata.abu$dij
out3 <- iNEXT3D(data[,1], diversity = 'FD', datatype = "abundance", nboot = 0, FDdistM = distM, FDtype = 'tau_values')
ggiNEXT3D(out3)

# diversity = 'FD' & FDtype = 'AUC'
data(FunDdata.abu)
data <- FunDdata.abu$data
distM <-  FunDdata.abu$dij
out4 <- iNEXT3D(data = data[,2], diversity = 'FD', datatype = "abundance", nboot = 0, FDdistM = distM)
ggiNEXT3D(out4)

## example for incidence-based data
# diversity = 'TD'
data(ant)
t <- round(seq(10, 500, length.out = 20))
out5 <- iNEXT3D(ant$h500m, diversity = 'TD', q = 1, datatype = "incidence_freq", size = t)
ggiNEXT3D(out5)

# diversity = 'PD'
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
nT <- data.inc$nT
out6 <- iNEXT3D(data, diversity = 'PD', q = c(0, 1, 2), datatype = "incidence_raw", nT = nT, PDtree = tree)
ggiNEXT3D(out6, facet.var = "Order.q", color.var = "Assemblage")

# diversity = 'FD' & FDtype = 'tau_values'
data(FunDdata.inc)
data <- FunDdata.inc$data
distM <-  FunDdata.inc$dij
out7 <- iNEXT3D(data, diversity = 'FD', datatype = "incidence_freq", FDdistM = distM, FDtype = 'tau_values')
ggiNEXT3D(out7)

# diversity = 'FD' & FDtype = 'AUC'
data(FunDdata.inc)
data <- FunDdata.inc$data
distM <-  FunDdata.inc$dij
out8 <- iNEXT3D(data, diversity = 'FD', datatype = "incidence_freq", nboot = 0, FDdistM = distM)
ggiNEXT3D(out8)

}
