% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggObsAsy3D}
\alias{ggObsAsy3D}
\title{ggplot2 extension for plotting q-profile, time-profile, and tau-profile}
\usage{
ggObsAsy3D(output, profile = "q")
}
\arguments{
\item{output}{the output of the function \code{ObsAsy3D}.\cr}

\item{profile}{a selection of profile versus to diversity. User can choose \code{'q'}, \code{'time'}, and \code{'tau'}. Default is \code{'q'} profile. \code{'time'} profile for only when \code{diversity = "PD"}. \code{'tau'} profile for only when \code{diversity = "FD"} and \code{FDtype = "tau_values"}.\cr}
}
\value{
a figure of asymptotic and observed diversity in q-profile, time-profile, or tau-profile.\cr\cr
}
\description{
\code{ggObsAsy3D} the \code{\link[ggplot2]{ggplot}} extension for \code{\link{ObsAsy3D}} object to plot q-profile, time-profile, and tau-profile based on the output of \code{ObsAsy3D} using the ggplot2 package.\cr
Note, it will only show the confidence interval of 'Asymptotic' when setting \code{method = "both"} in \code{ObsAsy3D}.
}
\examples{
\donttest{
# diversity = 'TD' for abundance-based data
data(Brazil_rainforest_data)
output1 <- ObsAsy3D(Brazil_rainforest_data$data, diversity = 'TD', datatype = "abundance")
ggObsAsy3D(output1)


# diversity = 'PD' for abundance-based data
data(Brazil_rainforest_data)
data <- Brazil_rainforest_data$data
tree <- Brazil_rainforest_data$tree
output2 <- ObsAsy3D(data, diversity = 'PD', q = seq(0, 2, by = 0.25), datatype = "abundance", 
                nboot = 20, PDtree = tree, PDtype = "meanPD")
ggObsAsy3D(output2, profile = "q")


# diversity = 'FD' & FDtype = 'tau_values' for abundance-based data
data(Brazil_rainforest_data)
data <- Brazil_rainforest_data$data
distM <- Brazil_rainforest_data$dist
output3 <- ObsAsy3D(data, diversity = 'FD', q = c(0, 1, 2), datatype = "abundance", nboot = 0, 
                FDtau = seq(0, 1, 0.1), FDdistM = distM, FDtype = 'tau_values')
ggObsAsy3D(output3, profile = "tau")


# diversity = 'FD' & FDtype = 'AUC' for abundance-based data
data(Brazil_rainforest_data)
data <- Brazil_rainforest_data$data
distM <- Brazil_rainforest_data$dist
output4 <- ObsAsy3D(data, diversity = 'FD', q = seq(0, 2, 0.5), datatype = "abundance", nboot = 10, 
                    FDdistM = distM, FDtype = 'AUC')
ggObsAsy3D(output4)


# diversity = 'TD' for incidence-based data
data(fish_incidence_data)
output5 <- ObsAsy3D(fish_incidence_data$data, diversity = 'TD', datatype = "incidence_raw")
ggObsAsy3D(output5)


# diversity = 'PD' for incidence-based data
data(fish_incidence_data)
data <- fish_incidence_data$data
tree <- fish_incidence_data$tree
output6 <- ObsAsy3D(data, diversity = 'PD', q = c(0, 1, 2), datatype = "incidence_raw", nboot = 20, 
                    PDtree = tree, PDreftime = seq(0.1, 1, length.out = 40), method = "Observed")
ggObsAsy3D(output6, profile = "time")


# diversity = 'FD' & FDtype = 'tau_values' for incidence-based data
data(fish_incidence_data)
data <- fish_incidence_data$data
distM <- fish_incidence_data$dist
output7 <- ObsAsy3D(data, diversity = 'FD', datatype = "incidence_raw", nboot = 20, 
                    FDdistM = distM, FDtype = 'tau_values')
ggObsAsy3D(output7)


# diversity = 'FD' & FDtype = 'AUC' for incidence-based data
data(fish_incidence_data)
data <- fish_incidence_data$data
distM <- fish_incidence_data$dist
output8 <- ObsAsy3D(data, diversity = 'FD', datatype = "incidence_raw", nboot = 20, 
                    FDdistM = distM, FDtype = 'AUC')
ggObsAsy3D(output8)
}


}
