% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggObsAsy3D}
\alias{ggObsAsy3D}
\title{ggplot2 extension for plotting q-profile, time-profile, and tau-profile}
\usage{
ggObsAsy3D(output, profile = "q")
}
\arguments{
\item{output}{the output of the function \code{ObsAsy3D}.\cr}

\item{profile}{a selection of profile versus to diversity. User can choose \code{'q'}, \code{'time'}, and \code{'tau'}. Default is \code{'q'} profile. \code{'time'} profile for only when \code{diversity = "PD"}. \code{'tau'} profile for only when \code{diversity = "FD"} and \code{FDtype = "tau_values"}.\cr}
}
\value{
a figure of asymptotic and observed diversity in q-profile, time-profile, or tau-profile.\cr\cr
}
\description{
\code{ggObsAsy3D} is a \code{ggplot} extension for \code{ObsAsy3D} object to plot q-profile, time-profile, and tau-profile based on the output of \code{ObsAsy3D} using the ggplot2 package.\cr
Note, it will only show the confidence interval of \code{'Asymptotic'} when setting \code{method = "both"} in \code{ObsAsy3D}.
}
\examples{
\donttest{
# diversity = 'TD' for abundance data
data(Brazil_rainforest_abun_data)
output_ObsAsy_TD_abun <- ObsAsy3D(Brazil_rainforest_abun_data, diversity = 'TD', 
                                  datatype = "abundance")
ggObsAsy3D(output_ObsAsy_TD_abun)


# diversity = 'PD' for abundance data
data(Brazil_rainforest_abun_data)
data(Brazil_rainforest_phylo_tree)
data <- Brazil_rainforest_abun_data
tree <- Brazil_rainforest_phylo_tree
output_ObsAsy_PD_abun <- ObsAsy3D(data, diversity = 'PD', q = seq(0, 2, by = 0.5), 
                                  datatype = "abundance", nboot = 20, PDtree = tree, 
                                  PDtype = "meanPD")
ggObsAsy3D(output_ObsAsy_PD_abun, profile = "q")


# diversity = 'FD' for abundance data
data(Brazil_rainforest_abun_data)
data(Brazil_rainforest_distance_matrix)
data <- Brazil_rainforest_abun_data
distM <- Brazil_rainforest_distance_matrix
output_ObsAsy_FD_abun <- ObsAsy3D(data, diversity = 'FD', q = seq(0, 2, 0.5), 
                                  datatype = "abundance", nboot = 10, 
                                  FDdistM = distM, FDtype = 'AUC')
ggObsAsy3D(output_ObsAsy_FD_abun)


# diversity = 'TD' for incidence data
data(Fish_incidence_data)
output_ObsAsy_TD_inci <- ObsAsy3D(Fish_incidence_data, diversity = 'TD', datatype = "incidence_raw")
ggObsAsy3D(output_ObsAsy_TD_inci)


# diversity = 'PD' for incidence data
data(Fish_incidence_data)
data(Fish_phylo_tree)
data <- Fish_incidence_data
tree <- Fish_phylo_tree
output_ObsAsy_PD_inci <- ObsAsy3D(data, diversity = 'PD', q = c(0, 1, 2), 
                                  datatype = "incidence_raw", nboot = 20, PDtree = tree, 
                                  PDreftime = seq(0.1, 1, length.out = 40), method = "Observed")
ggObsAsy3D(output_ObsAsy_PD_inci, profile = "time")


# diversity = 'FD' for incidence data
data(Fish_incidence_data)
data(Fish_distance_matrix)
data <- Fish_incidence_data
distM <- Fish_distance_matrix
output_ObsAsy_FD_inci <- ObsAsy3D(data, diversity = 'FD', datatype = "incidence_raw", nboot = 20, 
                                  FDdistM = distM, FDtype = 'AUC')
ggObsAsy3D(output_ObsAsy_FD_inci)
}


}
