% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{iNEXT3D}
\alias{iNEXT3D}
\title{iNterpolation and EXTrapolation of Hill number}
\usage{
iNEXT3D(
  data,
  class,
  q = c(0, 1, 2),
  datatype = "abundance",
  size = NULL,
  endpoint = NULL,
  knots = 40,
  conf = 0.95,
  nboot = 50,
  tree = NULL,
  nT = NULL,
  reftime = NULL,
  PDtype = "PD",
  distM,
  threshold = NULL
)
}
\arguments{
\item{data}{a matrix, data.frame (species by sites), or list of species abundances or incidence frequencies. If \code{datatype = "incidence_freq"}, then the first entry of the input data must be total number of sampling units in each column or list.}

\item{q}{a numerical vector of the order of Hill number.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).}

\item{size}{an integer vector of sample sizes (number of individuals or sampling units) for which diversity estimates will be computed. 
If NULL, then diversity estimates will be computed for those sample sizes determined by the specified/default \code{endpoint} and \code{knots} .}

\item{endpoint}{an integer specifying the sample size that is the \code{endpoint} for rarefaction/extrapolation. 
If NULL, then \code{endpoint} \code{=} double reference sample size.}

\item{knots}{an integer specifying the number of equally-spaced \code{knots} (say K, default is 40) between size 1 and the \code{endpoint};
each knot represents a particular sample size for which diversity estimate will be calculated.  
If the \code{endpoint} is smaller than the reference sample size, then \code{iNEXT3D()} computes only the rarefaction esimates for approximately K evenly spaced \code{knots}. 
If the \code{endpoint} is larger than the reference sample size, then \code{iNEXT3D()} computes rarefaction estimates for approximately K/2 evenly spaced \code{knots} between sample size 1 and the reference sample size, and computes extrapolation estimates for approximately K/2 evenly spaced \code{knots} between the reference sample size and the \code{endpoint}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95.}

\item{nboot}{an integer specifying the number of replications.}

\item{tree}{a phylo object describing the phylogenetic tree in Newick format for all observed species in the pooled assemblage. It is necessary when \code{class = 'PD'}.}

\item{nT}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the number of sampling units in each assemblage.
If \code{names(nT) = NULL}, then assemblage are automatically named as "assemblage1", "assemblage2",..., etc.
It is necessary when \code{class = 'PD'} and \code{datatype = "incidence_raw"}.}

\item{reftime}{is set to be the tree depth of the phylogenetic tree, which is spanned by all the observed species in
the pooled assemblage. Default is \code{NULL}. It will be use when \code{class = 'PD'}.}

\item{PDtype}{desired phylogenetic diversity type: \code{PDtype = "PD"} for Chao et al. (2010) phylogenetic diversity
and \code{PDtype = "meanPD"} for mean phylogenetic diversity (phylogenetic Hill number). It will be use when \code{class = 'PD'}. Default is \code{"PD"}.}

\item{distM}{a pair wise distance matrix for all pairs of observed species in the pooled assemblage. It will be use when \code{class = 'FD' or 'AUC'}.}

\item{threshold}{a sequence between 0 and 1 specifying tau. If \code{NULL}, \code{threshold = } dmean. Default is \code{NULL}. It will be use when \code{class = 'FD'}.}
}
\value{
a list of three objects: \code{$DataInfo} for summarizing data information; 
\code{$iNextEst} for showing diversity estimates for rarefied and extrapolated samples along with related statistics;
and \code{$AsyEst} for showing asymptotic diversity estimates along with related statistics.
}
\description{
\code{iNEXT3D}: Interpolation and extrapolation of Hill number with order q
}
\examples{
## example for abundance based data (list of vector)
# class = 'TD'
data(spider)
out1 <- iNEXT3D(spider, class = 'TD', q = c(0,1,2), datatype = "abundance")
out1$DataInfo # showing basic data information.
out1$AsyEst # showing asymptotic diversity estimates.
out1$iNextEst # showing diversity estimates with rarefied and extrapolated.

# class = 'PD'
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out2 <- iNEXT3D(data, class = 'PD', tree = tree, datatype = "abundance", q = c(0, 1, 2), nboot = 30)
out2

# class = 'FD'
data(FunDdata.abu)
data <- FunDdata.abu$data
dij <-  FunDdata.abu$dij
out3 <- iNEXT3D(data[,1], class = 'FD', distM = dij, datatype = "abundance", nboot = 0)
out3

# class = 'AUC'
data(FunDdata.abu)
data <- FunDdata.abu$data
dij <-  FunDdata.abu$dij
out4 <- iNEXT3D(data = data[,2], class = 'AUC', distM = dij, datatype = "abundance", nboot = 0)
out4

## example for incidence-based data
# class = 'TD'
data(ant)
t <- round(seq(10, 500, length.out = 20))
out5 <- iNEXT3D(ant$h500m, class = 'TD', q = 1, datatype = "incidence_freq", size = t)
out5

# class = 'PD'
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
nT <- data.inc$nT
out6 <- iNEXT3D(data, class = 'PD', nT = nT, datatype = "incidence_raw", tree = tree, q = c(0, 1, 2))
out6

# class = 'FD'
data(FunDdata.inc)
data <- FunDdata.inc$data
dij <-  FunDdata.inc$dij
out7 <- iNEXT3D(data, class = 'FD', distM = dij, datatype = "incidence_freq")
out7

# class = 'AUC'
data(FunDdata.inc)
data <- FunDdata.inc$data
dij <-  FunDdata.inc$dij
out8 <- iNEXT3D(data, class = 'AUC', distM = dij, datatype = "incidence_freq", nboot = 0)
out8

}
