% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ObsAsy3D}
\alias{ObsAsy3D}
\title{Asymptotic diversity and observed diversity of order q}
\usage{
ObsAsy3D(
  data,
  diversity = "TD",
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 50,
  conf = 0.95,
  nT = NULL,
  method = c("Asymptotic", "Observed"),
  PDtree,
  PDreftime = NULL,
  PDtype = "meanPD",
  FDdistM,
  FDtype = "AUC",
  FDtau = NULL,
  FDcut_number = 50
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a vector of species abundances (for a single assemblage), matrix/data.frame (species by assemblages), or a list of species abundance vectors. \cr
(b) For \code{datatype = "incidence_raw"}, data can be input as a list of matrix/data.frame (species by sampling units); data can also be input as a matrix/data.frame by merging all sampling units across assemblages based on species identity; in this case, the number of sampling units (nT, see below) must be input.}

\item{diversity}{selection of diversity type: \code{'TD'} = Taxonomic diversity, \code{'PD'} = Phylogenetic diversity, and \code{'FD'} = Functional diversity.}

\item{q}{a numerical vector specifying the diversity orders. Default is seq(0, 2, by = 0.2).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being 0 (non-detection) or 1 (detection).}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}

\item{nT}{(required only when \code{datatype = "incidence_raw"} and input data is matrix/data.frame) a vector of nonnegative integers specifying the number of sampling units in each assemblage. If assemblage names are not specified, then assemblages are automatically named as "assemblage1", "assemblage2",..., etc.}

\item{method}{computing type. Select 'Asymptotic' or 'Observed'.}

\item{PDtree}{(required argument only for \code{diversity = "PD"}), a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{PDreftime}{(argument only for \code{diversity = "PD"}), a vector of numerical values specifying reference times for PD. Default is \code{NULL} (i.e., the age of the root of PDtree).}

\item{PDtype}{(argument only for \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"} (effective total branch length) or \code{PDtype = "meanPD"} (effective number of equally divergent lineages). Default is \code{"meanPD"}, where \code{meanPD = PD/tree depth}.}

\item{FDdistM}{(required argument only for \code{diversity = "FD"}), a species pairwise distance matrix for all species in the pooled assemblage.}

\item{FDtype}{(argument only for \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_values"} for FD under specified threshold values, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(argument only for \code{diversity = "FD"} and \code{FDtype = "tau_values"}), a numerical vector between 0 and 1 specifying tau values (threshold levels). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}

\item{FDcut_number}{(argument only for \code{diversity = "FD"} and \code{FDtype = "AUC"}), a numeric number to cut [0, 1] interval into equal-spaced sub-intervals to obtain the AUC value by integrating the tau-profile. Equivalently, the number of tau values that will be considered to compute the integrated AUC value. Default is FDcut_number = 50 A larger value can be set to obtain more accurate AUC value.}
}
\value{
a table of diversity table including the following arguments: 
\item{Assemblage}{the assemblage name.}
\item{Order.q}{the diversity order of q.}
\item{qTD, qPD, qFD}{the estimated asymptotic diversity or observed diversity of order q.} 
\item{s.e.}{standard error of diversity.}
\item{qTD.LCL, qPD.LCL, qFD.LCL and qTD.UCL, qPD.UCL, qFD.UCL}{the bootstrap lower and upper confidence limits for the diversity.}
\item{Method}{"Asymptotic" means asymptotic diversity and "Observed" means observed diversity.}
\item{Reftime}{the reference times for PD.}
\item{Type}{"PD" (effective total branch length) or "meanPD" (effective number of equally divergent lineages) for PD.}
\item{Tau}{the threshold of functional distinctiveness between any two species for FD (under FDtype = tau_values).}
}
\description{
\code{ObsAsy3D}: The estimated asymptotic and observed diversity of order q
}
\examples{
\donttest{
# diversity = 'TD' for abundance-based data
data(Brazil_rainforest_abun_data)
output1 <- ObsAsy3D(Brazil_rainforest_abun_data, diversity = 'TD', datatype = "abundance")
output1


# diversity = 'PD' for abundance-based data
data(Brazil_rainforest_abun_data)
data(Brazil_rainforest_phylo_tree)
data <- Brazil_rainforest_abun_data
tree <- Brazil_rainforest_phylo_tree
output2 <- ObsAsy3D(data, diversity = 'PD', q = seq(0, 2, by = 0.25), datatype = "abundance", 
                    nboot = 20, PDtree = tree)
output2


# diversity = 'FD' & FDtype = 'tau_values' for abundance-based data
data(Brazil_rainforest_abun_data)
data(Brazil_rainforest_distance_matrix)
data <- Brazil_rainforest_abun_data
distM <- Brazil_rainforest_distance_matrix
output3 <- ObsAsy3D(data, diversity = 'FD', datatype = "abundance", 
                    nboot = 20, FDdistM = distM, FDtype = 'tau_values')
output3


# diversity = 'FD' & FDtype = 'AUC' for abundance-based data
data(Brazil_rainforest_abun_data)
data(Brazil_rainforest_distance_matrix)
data <- Brazil_rainforest_abun_data
distM <- Brazil_rainforest_distance_matrix
output4 <- ObsAsy3D(data, diversity = 'FD', q = seq(0, 2, 0.5), datatype = "abundance", 
                    nboot = 10, FDdistM = distM, FDtype = 'AUC')
output4


# diversity = 'TD' for incidence-based data
data(Fish_incidence_data)
output5 <- ObsAsy3D(Fish_incidence_data, diversity = 'TD', datatype = "incidence_raw")
output5


# diversity = 'PD' for incidence-based data
data(Fish_incidence_data)
data(Fish_phylo_tree)
data <- Fish_incidence_data
tree <- Fish_phylo_tree
output6 <- ObsAsy3D(data, diversity = 'PD', q = seq(0, 2, by = 0.25), datatype = "incidence_raw", 
                    nboot = 20, PDtree = tree)
output6


# diversity = 'FD' & FDtype = 'tau_values' for incidence-based data
data(Fish_incidence_data)
data(Fish_distance_matrix)
data <- Fish_incidence_data
distM <- Fish_distance_matrix
output7 <- ObsAsy3D(data, diversity = 'FD', datatype = "incidence_raw", 
                    nboot = 20, FDdistM = distM, FDtype = 'tau_values')
output7


# diversity = 'FD' & FDtype = 'AUC' for incidence-based data
data(Fish_incidence_data)
data(Fish_distance_matrix)
data <- Fish_incidence_data
distM <- Fish_distance_matrix
output8 <- ObsAsy3D(data, diversity = 'FD', datatype = "incidence_raw", 
                    nboot = 20, FDdistM = distM, FDtype = 'AUC')
output8
}


}
\references{
Chao, A., Henderson, P. A., Chiu, C.-H., Moyes, F., Hu, K.-H., Dornelas, M and Magurran, A. E. (2021). Measuring temporal change in alpha diversity: a framework integrating taxonomic, phylogenetic and functional diversity and the iNEXT.3D standardization. Methods in Ecology and Evolution, 12, 1926-1940.
}
