% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFun.R
\name{print.iNEXT3D}
\alias{print.iNEXT3D}
\title{Printing iNEXT3D object}
\usage{
\method{print}{iNEXT3D}(x, ...)
}
\arguments{
\item{x}{an \code{iNEXT3D} object computed by \code{iNEXT3D}.}

\item{...}{additional arguments.}
}
\value{
a list of three objects (see \code{iNEXT3D} for more details) with simplified outputs and notes.
}
\description{
\code{print.iNEXT3D}: Print method for objects inheriting from class "iNEXT3D"
}
